requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
pydantic==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.17
pillow==11.0.0
numpy==2.1.0
python-dotenv==1.0.1
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
